+++ lib/karafka/cli/info.rb
@@ -19,7 +19,7 @@ module Karafka
           "Number of threads: #{config.concurrency}",
           "Boot file: #{Karafka.boot_file}",
           "Environment: #{Karafka.env}",
-          "Kafka hosts: #{config.kafka.hosts}",
+          "Kafka seed brokers: #{config.kafka.seed_brokers}",
           "Redis: #{config.redis.to_h}"
         ]