+++ spec/lib/karafka/setup/configurators/water_drop_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe Karafka::Setup::Configurators::WaterDrop do
       Karafka::Setup::Config.config.class,
       concurrency: ::Karafka::App.config.concurrency,
       kafka: OpenStruct.new(
-        hosts: ::Karafka::App.config.kafka.hosts
+        seed_brokers: ::Karafka::App.config.kafka.seed_brokers
       ),
       # Instance double has a private method called timeout, that's why we use
       # openstruct here
@@ -28,7 +28,7 @@ RSpec.describe Karafka::Setup::Configurators::WaterDrop do
       expect(WaterDrop.config.send_messages).to eq true
       expect(WaterDrop.config.connection_pool_size).to eq config.connection_pool.size
       expect(WaterDrop.config.connection_pool_timeout).to eq config.connection_pool.timeout
-      expect(WaterDrop.config.kafka.hosts).to eq config.kafka.hosts
+      expect(WaterDrop.config.kafka.hosts).to eq config.kafka.seed_brokers
       expect(WaterDrop.config.raise_on_failure).to eq true
     end
   end